+++ lib/kitabu/exporter.rb
@@ -21,12 +21,15 @@ module Kitabu
       helper = root_dir.join("config/helper.rb")
       load(helper) if helper.exist?
 
+      FileUtils.rm_rf root_dir.join("output").to_s
+
       export_pdf = [nil, "pdf"].include?(options[:only])
       export_epub = [nil, "mobi", "epub"].include?(options[:only])
       export_mobi = [nil, "mobi"].include?(options[:only])
       export_txt = [nil, "txt"].include?(options[:only])
 
       exported = []
+      exported << Parser::HTML.parse(root_dir)
       exported << Parser::HTML.parse(root_dir)
       exported << Parser::PDF.parse(root_dir) if export_pdf && Dependency.prince?
       exported << Parser::Epub.parse(root_dir) if export_epub