+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Remove support for other file types like textile and html.
 - Remove support for pygments.rb highlighter; now using Rouge.
 - Add support for .erb files; helpers can be defined within the `Kitabu::Helpers` module.
+- Add support for media format stylesheets (print, pdf, html, epub).